+++ src/mink/__init__.py
@@ -37,6 +37,7 @@ from .tasks import DofFreezingTask as DofFreezingTask
 from .tasks import EqualityConstraintTask as EqualityConstraintTask
 from .tasks import FrameTask as FrameTask
 from .tasks import KineticEnergyRegularizationTask as KineticEnergyRegularizationTask
+from .tasks import LookAtTask as LookAtTask
 from .tasks import Objective as Objective
 from .tasks import PostureTask as PostureTask
 from .tasks import RelativeFrameTask as RelativeFrameTask