+++ src/mink/tasks/__init__.py
@@ -8,6 +8,7 @@ from .frame_task import FrameTask as FrameTask
 from .kinetic_energy_regularization_task import (
     KineticEnergyRegularizationTask as KineticEnergyRegularizationTask,
 )
+from .look_at_task import LookAtTask as LookAtTask
 from .posture_task import PostureTask as PostureTask
 from .relative_frame_task import RelativeFrameTask as RelativeFrameTask
 from .task import BaseTask as BaseTask